split genai.go into semconv.go + span.go, rename test
