+++ plugins/otel/genai/semconv.go
@@ -22,11 +22,6 @@
 // See: https://opentelemetry.io/docs/specs/semconv/gen-ai/
 package genai
 
-import (
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/trace"
-)
-
 // Attribute keys following OpenTelemetry Gen AI semantic conventions.
 const (
 	AttrGenAIOperationName             = "gen_ai.operation.name"
@@ -60,87 +55,3 @@ const (
 	OperationChat        = "chat"
 	OperationToolCall    = "execute_tool"
 )
-
-// ModelCallAttrs holds fields extracted from an LLM request/response pair.
-// Build this from whatever source you have (gjson, struct, etc.) and pass
-// it to StampModelCallSpan to produce spec-compliant span attributes.
-type ModelCallAttrs struct {
-	// Provider is the LLM provider name (e.g. "openai", "anthropic", "google", "bedrock").
-	Provider string
-
-	// RequestModel is the model name from the request body.
-	RequestModel string
-
-	// ResponseModel is the model name from the response body (may differ from request).
-	ResponseModel string
-
-	// ResponseID is the provider's response identifier.
-	ResponseID string
-
-	// FinishReason is the normalized finish reason (use OTel enum values:
-	// "stop", "length", "content_filter", "tool_call", "error").
-	FinishReason string
-
-	// Token usage.
-	InputTokens  int
-	OutputTokens int
-	CachedTokens int
-}
-
-// StampModelCallSpan sets gen_ai.* attributes on an existing span.
-// The output matches plugins/otel model.go exactly so that spans from
-// the AI Gateway are indistinguishable from spans produced by ai-sdk-go.
-func StampModelCallSpan(span trace.Span, a *ModelCallAttrs) {
-	if a == nil {
-		return
-	}
-
-	// OTel Gen AI semconv has no gen_ai.response.model attribute.
-	// Fall back to ResponseModel when RequestModel is empty so that
-	// gen_ai.request.model is always populated when possible.
-	model := a.RequestModel
-	if model == "" {
-		model = a.ResponseModel
-	}
-
-	attrs := []attribute.KeyValue{
-		attribute.String(AttrGenAIOperationName, OperationChat),
-	}
-
-	if model != "" {
-		attrs = append(attrs, attribute.String(AttrGenAIRequestModel, model))
-	}
-
-	if a.Provider != "" {
-		attrs = append(attrs, attribute.String(AttrGenAIProviderName, a.Provider))
-	}
-
-	if a.ResponseID != "" {
-		attrs = append(attrs, attribute.String(AttrGenAIResponseID, a.ResponseID))
-	}
-
-	if a.FinishReason != "" {
-		attrs = append(attrs, attribute.StringSlice(AttrGenAIResponseFinishReasons, []string{a.FinishReason}))
-	}
-
-	attrs = append(attrs,
-		attribute.Int(AttrGenAIUsageInputTokens, a.InputTokens),
-		attribute.Int(AttrGenAIUsageOutputTokens, a.OutputTokens),
-	)
-
-	if a.CachedTokens > 0 {
-		attrs = append(attrs, attribute.Int(AttrGenAIUsageCacheReadInputTokens, a.CachedTokens))
-	}
-
-	span.SetAttributes(attrs...)
-}
-
-// SpanName returns the canonical span name for a model call: "chat {model}".
-// If model is empty, returns just "chat".
-func SpanName(model string) string {
-	if model == "" {
-		return OperationChat
-	}
-
-	return OperationChat + " " + model
-}
