+++ plugins/otel/genai/span_test.go
